fix(stt): add module decls, clean imports, constants (#126)
- Add constants module declaration to lib.rs
- Move SttMetrics from processor.rs to types.rs (breaks circular dependency)
- Clean imports in helpers.rs (remove circular import, alphabetize)
- Replace magic numbers in processor.rs with named constants:
  - 16000 -> SAMPLE_RATE_HZ
  - 10 -> DEFAULT_BUFFER_DURATION_SECONDS
  - 100 -> LOGGING_INTERVAL_FRAMES
  - 5 -> SEND_TIMEOUT_SECONDS
- Comment out helpers module (requires coldvox_telemetry, should move to app crate)
